requirements.txt
anyio==3.6.2
click==8.1.3
dnspython==2.3.0
fastapi==0.93.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
motor==3.1.1
pydantic==1.10.6
pymongo==4.3.3
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.5.post1
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
